# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Polo Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clubs/{clubId}/teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List club teams
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Teams list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Team created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{teamId}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team details
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Team details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamDetail"
    put:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: Team updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
...
components:
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        name:
          type: string
        shortName:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        logoThumbUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        handicap:
          type: string
          nullable: true
        contactName:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - id
        - clubId
        - name
    TeamDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        name:
          type: string
        shortName:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        logoThumbUrl:
          type: string
          nullable: true
        primaryColor:
          type: string
          nullable: true
        handicap:
          type: string
          nullable: true
        contactName:
          type: string
          nullable: true
        contactPhone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        managerAssignment:
          $ref: "#/components/schemas/TeamManagerAssignment"
      required:
        - id
        - clubId
        - name
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        handicap:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        notes:
          type: string
        logoUrl:
          type: string
      required:
        - name
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        handicap:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        notes:
          type: string
        logoUrl:
          type: string
...
    TeamManagerAssignment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
          nullable: true
        teamId:
          type: string
          format: uuid
        tournamentId:
          type: string
          format: uuid
          nullable: true
        assignedBy:
          type: string
          format: uuid
          nullable: true
        inviteEmail:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending_invite, active, revoked]
        createdAt:
          type: string
          format: date-time
        acceptedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - teamId
        - status
...